+++ lib/pages/subscription_history/subscription_history_screen.dart
@@ -16,14 +16,16 @@ import 'package:stoyco_subscription/pages/subscription_history/screens/subscript
 /// {@endtemplate}
 class ResponsiveSubscriptionHistoryScreen extends StatelessWidget {
   /// Creates a [ResponsiveSubscriptionHistoryScreen].
-  const ResponsiveSubscriptionHistoryScreen({super.key});
+  const ResponsiveSubscriptionHistoryScreen({super.key, this.userId});
+
+  final String? userId;
 
   @override
   Widget build(BuildContext context) {
     if (!StoycoScreenSize.isPhone(context)) {
       return const Center(child: Text('En Desarrollo'));
     } else {
-      return const SubscriptionHistoryMobileScreen();
+      return SubscriptionHistoryMobileScreen(userId: userId ?? '');
     }
   }
 }